reorganize Search components, code cleanup
