+++ client/src/components/Search/ResultList.jsx
@@ -1,4 +1,6 @@
 var ResultItem = require('./ResultItem');
+
+
 var ResultList = React.createClass({
   render: function() {
     var resultList = this.props.list.map(function(company, index) {
